fix: extract types + drop selection resolution to O(V+S)
Boundaries: extract two self-contained types into their own files (cmux one-type-
per-file rule), reducing both oversized files by 200+ lines:
- FileExplorerStyle enum -> Sources/FileExplorerStyle.swift (FileExplorerStore 1794 -> 1573)
- FileExplorerCellView (+ its NSTextFieldDelegate extension) -> Sources/FileExplorerCellView.swift
  (FileExplorerView 2814 -> 2511)
Both wired into the Xcode project and the file-length budget updated to match.

Algorithmic complexity: applyStoredSelection now builds a path->row map once (O(V)) and
resolves multi-selection in O(V + S), instead of rescanning every visible row per selected
path (was O(S·V)). (clearReferences was already converted to O(N + M·depth).)
